backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
flask-migrate==4.1.0
SQLAlchemy==2.0.43
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10